♻️ refactor: interface 대신 type 사용 & 파일 위치 수정
