+++ FE/src/components/ModalOveray.tsx
@@ -1,0 +1,5 @@
+export default function Overay({ onClick }: { onClick: () => void }) {
+  return (
+    <div className='fixed inset-0 bg-black opacity-30' onClick={onClick}></div>
+  );
+}
