+++ FE/src/utils/socket.ts
@@ -1,0 +1,3 @@
+import { io } from 'socket.io-client';
+
+export const socket = io(import.meta.env.VITE_SOCKET_URL);
